+++ ui/app/src/models/GlobalState.ts
@@ -31,7 +31,6 @@ import { VersionsStateProps } from './Version';
 import QuickCreateItem from './content/QuickCreateItem';
 import { PathNavigatorStateProps } from '../components/PathNavigator';
 import { ContentItem } from './Item';
-import { CopyDialogStateProps } from '../components/CopyDialog/utils';
 import { PathSelectionDialogStateProps } from '../components/PathSelectionDialog/PathSelectionDialog';
 import { WidgetDescriptor } from './WidgetDescriptor';
 import { ItemMenuStateProps } from '../components/ItemActionsMenu';