+++ ui/app/src/CHANGELOG.md
@@ -64,6 +64,7 @@
   * [DependenciesDialog] Updated `dependencies` prop to be of type `ContentItem[] | LightItem[]` and `renderAction` prop to be of type `(item: ContentItem | LightItem) => ReactNode`.
   * [RenameContentDialogContainer] Updated `dependantItems` prop to be of type `LightItem[]`.
   * [RenameItemView] Updated `dependantItems` prop to be of type `LightItem[]`.
+  * Removed `CopyDialog` component.
 * [hooks]
   * Removed `useLogicResource` hook.
   * Removed `useSelectorResource` hook.
@@ -115,6 +116,7 @@
   * Updated `dependencies/fetchDependencies` to use API v2 (`/studio/api/2/dependency/{siteId}/publish_dependencies`)
   * Updated `dependencies/fetchSimpleDependencies` to use API v2 (`/studio/api/2/dependency/{siteId}/dependencies`)
   * Updated `dependencies/fetchDependant` to use API v2 (`/studio/api/2/dependency/{siteId}/dependent_items`)
+  * Removed `content/fetchLegacyItemsTree` service.
 * `PublishingItem` interface changes:
   * `approver` is now `reviewer`, of type Person.
   * `comment` is removed, and now there's `reviewerComment` and `submitterComment`.